requirements.txt
django
pillow
moviepy
numpy
requests
